refactor: 分离命令处理和管理
feat: 能不能好好说话 (WIP)

+++ src/main/kotlin/io/github/starwishsama/comet/api/command/CommandProps.kt
@@ -20,5 +20,5 @@ data class CommandProps(
     val permission: String,
     val level: UserLevel,
     val consumerType: CommandExecuteConsumerType = CommandExecuteConsumerType.COOLDOWN,
-    val consumePoint: Int = CometVariables.cfg.coolDownTime
+    val consumePoint: Double = CometVariables.cfg.coolDownTime.toDouble(),
 )
